Restore main
Function main in main.py was restored.
Program no longer freezes when there is no path between specified points. Instead, when maze is random, a new one is generated. If maze is from file, a ValueError is raised.
In random mazes, walls under points are replaced with blank cells.
In image generator, scale is mapped to an intiger.